requirements.txt
streamlit
Pillow
transformers
torch
langchain
langchain-community
langchain-core
langchain-ollama
